requirements.txt
appdirs==1.4.3
asgiref==3.3.1
asn1crypto==1.4.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.7.0rc1
distlib==0.3.0
distro==1.4.0
Django==3.1.3
django-braces==1.14.0
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.3
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.2
drf-oidc-auth==0.10.0
future==0.18.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
pycryptodomex==3.9.9
pyjwkest==1.4.2
PyJWT==1.7.1
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
urllib3==1.25.8
webencodings==0.5.1
